git est cassé ?

+++ robot_code/Path_planning/Path_planning.hh
@@ -19,7 +19,7 @@ typedef struct Path_planning
 {
 	double norm;  ///< current norm of the force vector field
 	double theta; ///< current angle of the force vector field
-	int v_max;    ///< max speed
+	double v_max;    ///< max speed
 
 } Path_planning;
